old/requirements.txt
digi-xbee
python-dotenv